adaptilearn/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
google-generativeai==0.4.1
PyPDF2==3.0.1
networkx==3.2.1
python-multipart==0.0.9
